test: 新增并重构数据库与搜索引擎连接器测试
1.  为OpenSearch、PostgreSQL、MariaDB、Oracle表连接器新增完整单元测试
2.  新增DSL条件转换测试用例覆盖全量过滤场景
3.  删除冗余重复的OpenSearch分组、全文检索测试文件
4.  补充全文搜索字段名相关测试验证
